+++ src/main/java/com/example/loadtest/domain/order/controller/OrderController.java
@@ -2,7 +2,9 @@ package com.example.loadtest.domain.order.controller;
 
 import com.example.loadtest.domain.order.dto.OrderRequestDTO;
 import com.example.loadtest.domain.order.service.OrderService;
+
 import lombok.RequiredArgsConstructor;
+
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -16,12 +18,12 @@ import java.net.URI;
 @RequiredArgsConstructor
 public class OrderController {
 
-    private final OrderService orderService;
+  private final OrderService orderService;
 
-    @PostMapping
-    public ResponseEntity<Void> createOrder(@RequestBody OrderRequestDTO requestDTO) {
-        Long orderId = orderService.createOrder(requestDTO);
-        URI location = URI.create("/api/orders/" + orderId);
-        return ResponseEntity.created(location).build();
-    }
+  @PostMapping
+  public ResponseEntity<Void> createOrder(@RequestBody OrderRequestDTO requestDTO) {
+    Long orderId = orderService.createOrder(requestDTO);
+    URI location = URI.create("/api/orders/" + orderId);
+    return ResponseEntity.created(location).build();
+  }
 }
