+++ src/main/java/com/example/loadtest/domain/order/dto/OrderRequestDTO.java
@@ -6,11 +6,17 @@ import java.util.List;
 
 @Getter
 public class OrderRequestDTO {
-        private Long userId;
-        private Long storeId;
-        private String mainMenuName;
-        private List<OrderItemDTO> orderItems;
-        private PaymentDTO payment;
-        private DeliveryInfoDTO delivery;
+
+  private Long userId;
+
+  private Long storeId;
+
+  private String mainMenuName;
+
+  private List<OrderItemDTO> orderItems;
+
+  private PaymentDTO payment;
+
+  private DeliveryInfoDTO delivery;
 
 }
