+++ src/main/java/com/example/loadtest/domain/order/dto/OrderResponseDTO.java
@@ -1,12 +1,17 @@
 package com.example.loadtest.domain.order.dto;
 
-import java.time.LocalDateTime;
+import java.time.ZonedDateTime;
 import java.util.List;
 
 public class OrderResponseDTO {
-    private Long orderId;
-    private String storeName;
-    private LocalDateTime orderTime;
-    private List<OrderItemDTO> items;
-    private int totalPaid;
+
+  private Long orderId;
+
+  private String storeName;
+
+  private ZonedDateTime orderTime;
+
+  private List<OrderItemDTO> items;
+
+  private int totalPaid;
 }
