+++ src/main/java/com/example/loadtest/domain/order/dto/OrderSummaryDTO.java
@@ -2,7 +2,8 @@ package com.example.loadtest.domain.order.dto;
 
 public class OrderSummaryDTO {
 
-    // 그냥 DTO 만들지 말고 String을 뱉는게 낫겠다..
-    private String mainMenuName;
-    private int otherCnt;
+  // 그냥 DTO 만들지 말고 String을 뱉는게 낫겠다..
+  private String mainMenuName;
+
+  private int otherCnt;
 }
