+++ src/main/java/com/example/loadtest/domain/order/dto/PaymentDTO.java
@@ -4,9 +4,14 @@ import lombok.Getter;
 
 @Getter
 public class PaymentDTO {
-    private String paymentMethod;
-    private int menuTotal;
-    private int deliveryFee;
-    private int discountAmount;
-    private int totalPaid;
+
+  private String paymentMethod;
+
+  private int menuTotal;
+
+  private int deliveryFee;
+
+  private int discountAmount;
+
+  private int totalPaid;
 }
