+++ src/main/java/com/example/loadtest/domain/order/entity/DeliveryInfo.java
@@ -5,7 +5,7 @@ import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import java.time.LocalDateTime;
+import java.time.ZonedDateTime;
 
 @Entity
 @Table(name = "delivery_info")
@@ -13,39 +13,45 @@ import java.time.LocalDateTime;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class DeliveryInfo {
 
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  private Long id;
 
-    @OneToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "order_id")
-    private Order order;
+  @OneToOne(fetch = FetchType.LAZY)
+  @JoinColumn(name = "order_id")
+  private Order order;
 
-    private String phoneNumber;
-    private String deliveryAddress;
-    private String messageToRider;
-    private String messageToStore;
+  private String phoneNumber;
 
-    @Column(updatable = false)
-    private LocalDateTime createdAt;
+  private String deliveryAddress;
 
-    @PrePersist
-    public void prePersist() {
-        this.createdAt = LocalDateTime.now();
-    }
+  private String messageToRider;
 
-    //== 연관관계 편의 메서드 ==//
-    void assignToOrder(Order order) {
-        this.order = order;
-    }
+  private String messageToStore;
 
-    //== 생성 메서드 ==//
-    public static DeliveryInfo create(String phoneNumber, String deliveryAddress, String messageToRider, String messageToStore) {
-        DeliveryInfo deliveryInfo = new DeliveryInfo();
-        deliveryInfo.phoneNumber = phoneNumber;
-        deliveryInfo.deliveryAddress = deliveryAddress;
-        deliveryInfo.messageToRider = messageToRider;
-        deliveryInfo.messageToStore = messageToStore;
-        return deliveryInfo;
-    }
+  @Column(updatable = false)
+  private ZonedDateTime createdAt;
+
+  @PrePersist
+  public void prePersist() {
+    this.createdAt = ZonedDateTime.now();
+  }
+
+
+  //== 연관관계 편의 메서드 ==//
+  // package-private 접근제한자. Order 엔티티를 통해서만 호출되어야 한다.
+  void assignToOrder(Order order) {
+    this.order = order;
+  }
+
+  //== 생성 메서드 ==//
+  public static DeliveryInfo create(String phoneNumber, String deliveryAddress,
+      String messageToRider, String messageToStore) {
+    DeliveryInfo deliveryInfo = new DeliveryInfo();
+    deliveryInfo.phoneNumber = phoneNumber;
+    deliveryInfo.deliveryAddress = deliveryAddress;
+    deliveryInfo.messageToRider = messageToRider;
+    deliveryInfo.messageToStore = messageToStore;
+    return deliveryInfo;
+  }
 }
